refactor(session): extract TestHarnessMixin to _session_testharness.py
Behavior-preserving move of _load_test_*, _test_nav_drag,
_load_test_vectors, _run_test_orientation into a mixin.
